Updated version of c-dict.
This updates c-dict to be up-to-date w.r.t. other changes in the
source, especially with changes in host/direct.

Additionally, this adds primitive defintions for all of the
primitives under host/direct by copying the associated direct-mode
sources, with modifications to make it work.  The resulting c-dict
output isn't _quite_ correct, because of the use of things like
<C> and [compile] in the directmode source.