Force dictionary building to use the Forth QUIT.
This invokes QUIT as the first thing in the first file in the
file order, forcing the interpreter into the Forth implementation
found in stdforth.fth.

The main reason for this is to guarantee that the implementation in
stdforth.fth stays working.  It should also allow us to start
tearing down the C code for quit().