Delete the search order fromthe static dictionary.
The search order isn't used in the static dictionay, so it can be
deleted.  Rewrote the key primitives to instead allocate the needed
space using ALLOT.

Additionally, added exception checking for "search-order overflow"
(throw code -49) and "search-order underflow" (throw code -50)

At this point, the contents of the static dictionary are reduced to
1) HERE and FORTH-WORDLIST, plus 2) global variables DICT.input and
DICT.lineno.