Convert dictionary primitives from C to Forth.
This converts dictionary allocation primitives such as HERE and ALLOT
from C to Forth.

At this point, there are no more primitives in C that could reasonably
be in Forth.  The only C code left to get rid of is the meta-interpret
and direct-mode functions used during dictionary initialization at
startup.