Added a benchmark to measure C divide times.
The benchmark measures the cost of a / b and a % b each
individually, and then the cost when they're combined together.
Some hardware+compiler combinations may be able to optimize both
(a / b) and (a % b) down to a single instruction sequence that
calculates both values at once.  This benchmark can find whether
we're on such a platform.