Split forth.h into five sub-header files.
The five new files aren't truly independent; they must be included
in a specific order because of dependencies.  The intent (for now)
is that the code continues to include "forth.h", which includes the
pieces in the proper order.

I don't know for sure that I'll keep it this way.  For now, it's
useful to have the bits broken down into smaller chunks where the
dependencies are easier to see.