Rename selected names in xtoken.h.
There was a modest number of name changes to types defined in
xtoken.h.  Not that many names, but they were the most popular,
so pretty much every file got touched.  :-(

Added PRIM_HDLR() and DEFINER() macros to standardize declarations
and definitions of XT handler functions.  This will also add a bit
of security if I get tired of the new names, too. 8-)