Better internal checking of CATCH and THROW.
Adjusted return stack bounds checking by redefining the bottom of
the return stack as catch_rsp.  That means CHECK_RPOP() throws
"return stack underflow" if you attempt to pop past the current
catch point, even if there are still cells on the stack.  The
only way to get past catch_rsp (safely) is to call undo_catch()
when a) the rsp is pointing to catch_rsp, and b) catch_rsp isn't
the bottom of the stack.