Simplified error reporting yet again.
'lookup' already checks for error -16, so callers that say
'parse-name lookup' are protected.  After accounting for that,
there doesn't seem to be any call for shared words that detect
when to invoke certain errors; every case seems to be different.

Some errors are still thrown from more than one site:  This seems
to be because each site is qualitatively different in some way.
  * -13 throw happens in two distinct cases:
	after INTERPRET-UNKNOWN reports "I couldn't handle it".
	when POSTPONE or ' or similar words can't find a name.
  * -16 throw happens in three cases:
	when attempting to add an empty name to namespace
	when attempting to look up an empty name
	when char parses an empty name