Refactored target.fth slightly.
In preparation for the direct-mode compiler, we need all instruction
generation in the DIRECT vocabulary to know whether input strings
were returned by PARSE or PARSE-NAME (and therefore transient) or
allocated in non-transient memory (i.e. the dictionary).

With this change, whether a passed in string is transient or
non-transient is now a property of the instruction.  I think...