Remove bogus error check in LOOKUP.
The code for `LOOKUP` included a check for looking up a zero-length
name, and would execute `-16 THROW` if it occurred.  That seemed mostly
wrong: it's probably preferred that lookups return "not found" in that
case instead.  Moreover, it didn't catch all the cases that might occur;
at minimum, there needed to be a check somewhere like `WID-LOOKUP`.

This change removes the check.  That'll cause `-13 THROW` instead.
We'll still get `-16 THROW` in cases where we try to _create_ an empty
name, just not when we try to find one.