Move `bootstrap.meta` to `reference`.
The `bootstrap.meta` directory has been sitting around at top-level for
a while, but the only reason I keep it is for easy reference to certain
ideas it embodies.  But, it's in the way, and it's not really a
top-level thing at all, it's "reference".

So, move things around to reflect the reality.