Delete lookup() from C sources.
This deletes lookup() from C sources, and further isolates knowledge
of the structure of namespace to one C file.

This also deletes direct_test.c; the file wasn't really used and
no longer reflects any idea that I'm actively considering.  The
file was in the way because it depended on "names.h", which is
being deleted.