Add new `meta` dictionary builder.
It's almost the same as `direct`.  It can easily be changed to
create a new initdict.c, that I assume would be correct (that's
not tested).  There also also notes and random other files
representing a WIP.

The plan is to turn this into a new dictionary builder that liberates
the code from the remnants of C that we don't really need/want.