Integrate direct code with new meta-compiler
The forth initialization builder now uses the new integrated bits;
dependencies on direct mode execution are gone, as are dependencies
on C primitives for the builder.

Some bits from `direct` are still included directly into `meta` in
`file-order`, and the C code for the meta-builder is still present
(and consequently, linked in unnecessarily).  Additional cleanup is
needed to better integrate the things I'm calling `TARGET` from the
thing I call `META`.