Fix building for Linux.
The order of arguments on the command that links the
"bootstrap/forth" executable didn't work on Linux (specifically, a
reasonably up-to-date Debian distribution for Raspberry Pi).  Re-order
the command line so that "-lreadline" follows the object files, rather
than precedes them.

The change is also tested for macOS, and works with the Xcode-supplied
command line compiler.