Add comments identifying standards in terminal.c
Generally, when there are header includes for any facility not specified
by the ISO/IEC standard for C, I want to put in a comment identifying
the specification or other source for the facility.  In that case,
comnments are also added to identify ISO/IEC standard headers.

The bootstrap/terminal.c file depends on the GNU "readline" library, so
this adds the conventional comments.