Allow gen-dict to be invoked without naming the metacompiler.
Up to now, `gen-dict` has required one argument identifying the
location of the "file-order" for the code that builds initdict.c; this
allowed experimenting with different metacompilers.  At this point,
there's no alternative to the main metacompiler, so the argument is
redundant.

Update the script to make the metacompiler argument optional, defaulting
to "meta".