Re-land "Further refactoring of file loading"
I thought this was wrong the first time I committed it, but I was
mistaken.

Move the defintions of COMPILE-FILE:, COMPILE-META:, et al. out of
compile-file.fth to be inline in output-dict.fth.

Update the loading of the first two primitive source files in the
meta-dictionary to have their names listed in comments in the output
C source, just like all the others.