Add reference assembly code for AArch64.
The code is copied from the amd64 reference implementation.  The code
passes tests, but that only means three .s files have been converted;
the remainders are still unconverted AMD64 assembly.

The code is partly duplicative of the AMD64 code; work should be done to
merge/integrate, but that can wait for true integration with the C-based
VM.