Make .error be internal to QUIT.
Numerous internal definitions would call .error to print an indicator of
where in the terminal input buffer a particular error occurred.  That
wasn't strictly necessary: QUIT could do the processing just as well, and
on reflection, letting QUIT do the handling seems more appropriate.

So, change all the calls to `.error` to be just `throw`, and move the
terminal input buffer printing to QUIT.